ALOT of updates and changes.  You will need to run the update.php script and I DO NOT RECOMMEND this for production as I kill the video_rendering table and create a new table video_files.  I restructured how the videos are saved and processed.
Now have access to original video object in your theme.  $video->original
I don't overwrite the orginal file anymore but create a link to the converted file and the original by file id.  Now when uploading a video through the cck widget, there is no manuipulation of the file object.  When the video is rendered is when I replace the paths with the converted file, this will allow us to also prep for CDN's.  So the CDN will break in at the same point and make for seemless uninterrupted processing.  It will be easy to turn off conversions and revert back to the original file, and same for CDN, you will be able to turn it off and go back to the original.

I now have good cleanup for videos and thumbnails.  Before the thumbnails were left and never removed.  I have fixed that.  I have also fixed when you delete a video, it will delete the orginal and it will delete the converted video.

Alot of code changes so this could be buggy, but from testing so far it looks good and were on the right path.